#1213: created react hook function for reporting refund, created api function, created url
