+++ src/frontend/src/pages/FinancePage/ReportRefundModal.tsx
@@ -1,11 +1,15 @@
-import { Box, FormControl, InputAdornment } from '@mui/material';
+import { Box, FormControl } from '@mui/material';
 import NERFormModal from '../../components/NERFormModal';
 import { useForm } from 'react-hook-form';
 import AttachMoneyIcon from '@mui/icons-material/AttachMoney';
 import NERFailButton from '../../components/NERFailButton';
 import NERSuccessButton from '../../components/NERSuccessButton';
-import { ReportRefundInputs } from './ReportRefund';
 import ReactHookTextField from '../../components/ReactHookTextField';
+import { useHistory, useParams } from 'react-router-dom';
+import LoadingIndicator from '../../components/LoadingIndicator';
+import ErrorPage from '../ErrorPage';
+import { useToast } from '../../hooks/toasts.hooks';
+import { useReportRefund } from '../../hooks/finance.hooks';
 
 interface ReportRefundViewProps {
   modalShow: boolean;
@@ -55,4 +59,37 @@ const ReportRefundView: React.FC<ReportRefundViewProps> = ({ modalShow, onHide, 
   );
 };
 
-export default ReportRefundView;
+interface ReportRefundProps {
+  modalShow: boolean;
+  handleClose: () => void;
+}
+
+export interface ReportRefundInputs {
+  newAccountCreditAmount: number;
+}
+
+const ReportRefund: React.FC<ReportRefundProps> = ({ modalShow, handleClose }: ReportRefundProps) => {
+  const history = useHistory();
+  const toast = useToast();
+  const id = useParams<{ id: string }>().id;
+  const { isLoading, isError, error, mutateAsync } = useReportRefund(id);
+
+  const handleConfirm = async ({ newAccountCreditAmount }: ReportRefundInputs) => {
+    handleClose();
+    await mutateAsync(newAccountCreditAmount).catch((error) => {
+      if (error instanceof Error) {
+        toast.error(error.message);
+      }
+    });
+    history.goBack();
+    toast.success(`New Account Credit Amount #${newAccountCreditAmount} Reported Successfully`);
+  };
+
+  if (isLoading) return <LoadingIndicator />;
+
+  if (isError) return <ErrorPage message={error?.message} />;
+
+  return <ReportRefundView modalShow={modalShow} onHide={handleClose} onSubmit={handleConfirm} />;
+};
+
+export default ReportRefund;
